[refactor] 파일 위치 변경
error 폴더를 추가하고 그곳에 파일을 이동
